+++ biopipen/scripts/scrna/CellTypeAnnotation-sctype.R
@@ -1,18 +1,9 @@
-library(dplyr)
 library(HGNChelper)
-library(Seurat)
-library(rlang)
-library(biopipen.utils)
 
 {% include biopipen_dir + "/scripts/scrna/sctype.R" %}
 
-sobjfile = {{in.sobjfile | r}}
-outfile = {{out.outfile | r}}
 tissue = {{envs.sctype_tissue | r}}
 db = {{envs.sctype_db | r}}
-newcol = {{envs.newcol | r}}
-ident = {{envs.ident | r }}
-merge_same_labels = {{envs.merge | r}}
 
 if (is.null(db)) { stop("`envs.sctype_args.db` is not set") }
 
@@ -21,14 +12,17 @@ log <- get_logger()
 log$info("Reading Seurat object...")
 sobj = biopipen.utils::read_obj(sobjfile)
 ident <- ident %||% biopipen.utils::GetIdentityColumn(sobj)
-Idents(sobj) <- ident
+if (is.null(ident)) {
+    sobj@meta.data$Identity <- Idents(sobj)
+    ident <- "Identity"
+}
 
 # prepare gene sets
 log$info("Preparing gene sets...")
 gs_list = gene_sets_prepare(db, tissue)
 
 scRNAseqData = GetAssayData(sobj, layer = "scale.data")
-idents = as.character(unique(Idents(sobj)))
+idents = as.character(unique(sobj@meta.data[[ident]]))
 idents = idents[order(as.numeric(idents))]
 
 log$info("Working on different levels of cell type labels ...")
@@ -107,32 +101,11 @@ if (length(cell_types_list) == 1) {
 
 
 log$info("Renaming cell types...")
-ct_numbering = list()
-for (key in names(celltypes)) {
-    ct = celltypes[[key]]
-    ct_numbering[[ct]] = ct_numbering[[ct]] %||% 0
-    if (ct_numbering[[ct]] > 0) {
-        celltypes[[key]] = paste0(ct, ".", ct_numbering[[ct]])
-    }
-    ct_numbering[[ct]] = ct_numbering[[ct]] + 1
-}
-
-celltypes = as.list(celltypes)
-if (is.null(newcol)) {
-    sobj <- rename_idents(sobj, ident, celltypes)
-} else {
-    sobj@meta.data[[newcol]] = celltypes[as.character(Idents(sobj))]
-}
-celltypes$object = NULL
-gc()
-
-if (merge_same_labels) {
-    log$info("Merging clusters with the same labels...")
-    sobj <- merge_clusters_with_same_labels(sobj, newcol)
-    celltypes <- lapply(celltypes, function(ct) {
-        sub("\\.\\d+$", "", ct)
-    })
-}
+sobj <- biopipen.utils::RenameSeuratIdents(
+    sobj,
+    celltypes,
+    ident = ident, save_as = newcol, merge = merge, backup = backup_col
+)
 
 log$info("Saving the mappings ...")
 write.table(